pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-thermal_grid_rl_agent"
version = "0.1.0"
description = "Thermal Grid Rl Agent environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "stable-baselines3[extra]>=2.8.0",
    "gymnasium>=1.2.3",
    "torch>=2.1.0,<2.4.0",
    "shimmy>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m thermal_grid_rl_agent.server.app
server = "thermal_grid_rl_agent.server.app:main"

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.setuptools]
include-package-data = true
packages = ["thermal_grid_rl_agent", "thermal_grid_rl_agent.server"]
package-dir = { "thermal_grid_rl_agent" = ".", "thermal_grid_rl_agent.server" = "server" }
